Handle Gui events in separate module
Handle Gui events in separate module
